Added base MVC packages and some package renaming
