+++ src/main/java/com/fetch/takeHome/receiptProcessor/ReceiptProcessorApplication.java
@@ -1,4 +1,4 @@
-package com.fetch_take_home.receipt_processor;
+package com.fetch.takeHome.receiptProcessor;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
