+++ src/main/resources/application.properties
@@ -1,9 +1,8 @@
 spring.application.name=receipt-processor
 
 ########## DATABASE STUFF ##########
+#access DB via UI (http://localhost:8080/h2-console)
 spring.datasource.driverClassName=org.h2.Driver
-# name of DB in H2
 spring.datasource.url=jdbc:h2:mem:test
-# access DB via UI (http://localhost:8080/h2-console)
 spring.h2.console.enabled=true
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
